Added more models, fixed bug with last combined depth analysis image Fixed bvhanalysis integer size for huge models, and added option to not render bvhTree image fixed crash with to large texture index in modelloader for now using easy bvh bucket version (not sure if the spacial is usefull)

+++ src/accelerationStructure/nodeAnalysis.h
@@ -60,7 +60,7 @@ public:
 	}
 
 	//traverses tree from left to right
-	void analysis(std::vector<NodeAnalysis*>& leafNodes, std::vector<uint16_t>& treeDepth, std::vector<uint16_t>& childCount, std::vector<uint16_t>& primCount)
+	void analysis(std::vector<NodeAnalysis*>& leafNodes, std::vector<uint32_t>& treeDepth, std::vector<uint32_t>& childCount, std::vector<uint32_t>& primCount)
 	{
 		size_t cc = children.size();
 		size_t pc = primitiveCount;
